Ensure waitContext.waitTimeMs * 1000 doesn't overflow uint32
Just for pedant analyzer warning fix. Very unlikely that wait time is greater than 1.2 hours to overflow, but still.

+++ Scheduler/Source/MTScheduler.cpp
@@ -402,7 +402,7 @@ namespace MT
 
 		bool isTaskStealingDisabled = context.taskScheduler->IsTaskStealingDisabled(0);
 
-		int64 timeOut = GetTimeMicroSeconds() + (waitContext.waitTimeMs * 1000);
+		int64 timeOut = GetTimeMicroSeconds() + ((int64)waitContext.waitTimeMs * 1000);
 
 		SpinWait spinWait;
